Move `PersistedDht` from eth2-libp2p to network crate
