+++ beacon_node/eth2-libp2p/Cargo.toml
@@ -27,8 +27,6 @@ unsigned-varint = "0.2.3"
 lazy_static = "1.4.0"
 lighthouse_metrics = { path = "../../eth2/utils/lighthouse_metrics" }
 tokio-io-timeout = "0.3.1"
-store = { path = "../store" }
-rlp = "0.4.3"
 
 [dev-dependencies]
 slog-stdlog = "4.0.0"
