+++ beacon_node/network/Cargo.toml
@@ -24,3 +24,4 @@ smallvec = "0.6.11"
 # TODO: Remove rand crate for mainnet
 rand = "0.7.2"
 fnv = "1.0.6"
+rlp = "0.4.3"
